ArmVirtPkg: Save DT base address from X0 in PCD
Some virtual machine managers provide the base address of the DT
in memory in the X0 register. Save the DT Base address in the
PcdDeviceTreeInitialBaseAddress so that the firmware can use the
PCD to parse the DT and obtain the platform information subsequently.

This change also requires that the PcdDeviceTreeInitialBaseAddress
be a Dynamic PCD.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>