MdeModulePkg: Map persistent (NV) memory
Some platforms are able to preserve a memory range across
system resets. This memory can be used for the Non-Volatile
variables storage. The PcdEmuVariableNvStoreReserved is
used to specify this option.

This patch enables mapping of this Non-Volatile memory range
as runtime memory so that the variable storage is accessible
post ExitBootServices.

Also added PcdMapEmuVariableNvStoreReserved to select if the
memory region described by PcdMapEmuVariableNvStoreReserved
should be mapped by the Variable Emulation driver.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>