MdePkg: Add NULL implementation for PCILib
On some platforms the Serial 16550 UART is interfaced
over PCI. To support such platforms the Serial 16550
driver links with PciLib.

For platforms that do not interface the Serial 16550
UART over PCI, the driver still needs to link with a
PciLib library. Linking to the full implementation of
the PCI library may not be possible during the early
firmware boot stage.

To facilitate early firmware logs over the serial port
this patch introduces a NULL PCI library implementation.

Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>