ArmPlatformPkg: Use MMIO to read device memory
NorFlashDxe must use aligned MMIO accesses to
read data from flash as this is device memory.

The AlignedCopyMem() was used to copy the flash
data which prevented unaligned access to device
memory. However, the compiler could optimize the
code to generate pre/post indexed or LDP
operations. This is a problem for guest/virtual
firmware as the hypervisor code cannot get the
syndrome information for the trapped accesses.

Similarly, the GUIDS FwVolHeader->FileSystemGuid
and VariableStoreHeader->Signature in
ValidateFvHeader() are compared using
CompareGuid(). These GUIDs point to flash data
(which is device memory) and therefore need
aligned MMIO accesses.

To address the above issues, BaseMemoryLibMmio
library has been introduced to perform aligned
MMIO accesses.

This patch removes the usage of AlignedCopyMem()
and replaces it with CopyMem() with an expectation
that the BaseMemoryLibMmio will be linked
with NorFlashDxe.

Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>