DynamicTablesPkg: SSDT Serial Port Table generator
According to Arm Server Base Boot Requirements,
Platform Design Document version 1.2 revision D,
September 2, 2019, section '4.2.1.8 SPCR';
The SPCR console device must be included in the DSDT.

Additionally it is often desirable to describe the
serial ports available on a platform so that they
are available for use by a rich OS.

The SSDT Serial Port Table Generator uses the
configuration manager protocol to obtain information
for the Serial Ports on the platform and generates
SSDT tables describing the Serial Port information.

The serial ports are described using the
CM_ARM_SERIAL_PORT_INFO structure in the configuration
manager. The EArmObjSerialPortInfo ID is used to
represent a standard serial port while the SPCR port
is represented using EArmObjSerialConsolePortInfo.

At build time a standard SSDT Serial port ASL file
is compiled to generate an AML template blob. This
AML blob is packaged in the firmware.

The SSDT Serial Port Table generator:
 - gets the serial port hardware information
   from the configuration manager.
 - uses the AmlLib interfaces to parse the AML
   template BLOB and construct an AML tree.
 - uses the AmlLib Iteration APIs to iterate the
   tree and locate the resource data node that
   contains the placeholder for the base address
   and interrupt information for the serial port.
 - performs a fixup of the template AML code by
   updating the resource data elements in the AML
   tree with the serial port information obtained
   from the configuration manager.
 - serializes the AML tree to an output buffer. This
   output buffer effectively contains the fixed-up
   AML code, which is then installed as an ACPI
   SSDT table.

Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>