DynamicTablesPkg: Add AmlLib to enable AML parsing
ACPI tables are categorized as:
  - data tables (e.g MADT, GTDT, etc.) and
  - definition blocks (e.g. DSDT or SSDT).

The DSDT and SSDT tables are used to describe system
devices along with other control and power management
information. These tables are written using Acpi Source
Language (ASL). The ASL code is compiled using an ASL
compiler (e.g. Intel iASL compiler) to generate Acpi
Machine Language (AML) bytecode. This AML code is
effectively called 'definition blocks'.

The Dynamic Tables framework implements mechanisms to
build ACPI data tables at run-time. This involves
updating fields in data tables (which are typically
represented as C structures).
Since, definition blocks are represented using AML
grammar, run-time generation of definition blocks is
complex. A full blown solution would mean implementing
an ASL run-time compiler.

An alternate solution is to compile an ASL template
file to generate AML bytecode. This template AML
bytecode can be parsed at run-time and a fixup code
can update the required fields in the AML template.

Since, AML bytecode represents complex AML grammar,
an AmlLib library is introduced to assist parsing and
traversing of the AML bytecode at run-time.

The AmlLib library parses a definition block and
represents it as an AML tree. This tree representation
is based on the AML grammar defined by the ACPI 6.3
specification, section - 20 'ACPI Machine Language (AML)
Specification'.

AML objects, methods, data are represented as tree nodes.
Since the AML data is represented as tree nodes, it is
possible to traverse the tree, locate a node and modify
the node data. The tree can then be serialized to a
buffer (that represents the definition block). This
definition block containing the fixed up AML code
can be installed as an ACPI table (DSDT/SSDT).

The set of external Apis is currently limited to the
functions necessary to update a device's name, the value
of a "_UID" object, and the memory and interrupt number
stored in a "_CRS" node. This set of Apis can be extended
depended on later requirements.

Although the AmlLib performs checks to a reasonable
extent while modifying a definition block, these checks
may not cover all aspects due to the complexity of the
ASL/AML language. It is therefore recommended to review
any operation performed, and validate the generated
output.
Example: The serialized AML code could be validated by
 - saving it to a file and comparing with a reference
   output or
 - decompiling using the iASL compiler.

Signed-off-by: Pierre Gondois <pierre.gondois@arm.com>
Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>