MdePkg: Base Memory Lib instance using MMIO
Some device drivers perform copy operations on
device memory, e.g. device drivers for a Flash
device. On some architectures unaligned access
to device memory regions is not permitted. To
add to this if the device is virtualised then
there are further restrictions on the type of
load/store operations that can be performed
on the device memory regions, e.g. on AARCH64,
Pre/Post index or LDP operations cannot be
used, as a trap to EL2 does not provide the
syndrome information to the hypervisor.

To address these issues this patch introduces
BaseMemoryLibMmio library which provides an
implementation of Base memory library that
uses aligned MMIO accesses.

Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>