ArmPlatformPkg: Fix comparison of constants warning
The VS2017 compiler reports 'warning C6326: potential
comparison of a constant with another constant' when
a fixed PCD value is compared with a constant value.

The faulting code is as marked by '-->' below:

--> if (FixedPcdGet32 (PL011UartInteger) != 0) {
      Integer = FixedPcdGet32 (PL011UartInteger);
      Fractional = FixedPcdGet32 (PL011UartFractional);
    } else {
    ...

The macro FixedPcdGet32 (PL011UartInteger) evaluates
to a macro _PCD_VALUE_PL011UartInteger that is defined
by the build system to represent the UART Integer
value. Therefore, the VS2017 compiler reports the above
warning.

In this case the warning reported by the Visual Studio
compiler does not evaluate to an issue. However, it can
be useful to detect potential issues in other scenarios.
Other compilers may either be incapable of detecting and
reporting comparison with constant warnings or may be
good at reducing false positives. So, it is definitely
useful to keep this warning enabled, and disabling it
case by case is a suitable option.

Therefore, disable this warning for Visual studio
compilers using the pragma suppress directive that:
'Pushes the current state of the pragma on the stack,
disables the specified warning for the next line, and
then pops the warning stack so that the pragma state
is reset.'

Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>